Añadir pruebas unitarias para InvitacionRepositorio
Se añadieron pruebas unitarias en `InvitacionRepositorioTests.cs`
para validar las operaciones CRUD del repositorio de invitaciones
(`InvitacionRepositorioImpl`). Las pruebas incluyen casos para
agregar, actualizar, eliminar y obtener invitaciones, verificando
el manejo correcto de relaciones con entidades como `Gimnasio` y
`Socio`.

Se agregó el método `CreateInMemoryOptions` para configurar una
base de datos en memoria y facilitar las pruebas. También se
corrigió el tipo de dato de `gimnasioId` en el método
`ObtenerTodasAsync` de `IInvitacionRepositorio` y
`InvitacionRepositorioImpl`, cambiándolo de `int` a `long` para
mantener consistencia con el modelo.

Se incluyó el uso de `FluentAssertions` para mejorar la claridad
de las aserciones y se añadieron dependencias necesarias como
`Microsoft.EntityFrameworkCore` y `Xunit`.

+++ FitRank-API/Infrastructure/Interfaces/IInvitacionRepositorio.cs
@@ -11,7 +11,7 @@ namespace FitRank_API.Infrastructure.Interfaces
         Task<bool> Eliminar(long id);
         Task<Invitacion> ObtenerPorIdYEstadoAsync(long invitacionId, string v);
 
-        Task<List<Invitacion>> ObtenerTodasAsync(int gimnasioId);
+        Task<List<Invitacion>> ObtenerTodasAsync(long gimnasioId);
 
         Task<Invitacion?> ObtenerPorEmailAsync(string email);
